pyproject.toml
[tool.poetry]
name = "price-alarm"
version = "0.1.0"
description = "Un sistema inteligente de monitoreo de precios que rastrea productos en tiendas en línea y envía alertas por Telegram"
authors = ["juanbarco92 <juanbarco92@users.noreply.github.com>"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/juanbarco92/Price-Alarm"
keywords = ["price-monitoring", "web-scraping", "telegram-bot", "alkosto", "automation"]
packages = [
    {include = "shared"},
    {include = "app"},
    {include = "scraper"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Office/Business :: Financial",
]

[tool.poetry.dependencies]
python = "^3.10"
# Shared dependencies
pydantic = "^2.0.0"
sqlalchemy = "^2.0.0"
psycopg2-binary = "^2.9.0"
python-dotenv = "^1.0.0"
pyyaml = "^6.0"

[tool.poetry.group.app.dependencies]
# Web application dependencies
flask = "^3.0.0"
flask-cors = "^4.0.0"
flask-sqlalchemy = "^3.0.0"
gunicorn = "^21.0.0"
jinja2 = "^3.1.0"

[tool.poetry.group.scraper.dependencies]
# Scraper dependencies
playwright = "^1.40.0"
python-telegram-bot = "^20.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.0"

[tool.poetry.group.dev.dependencies]
# Development dependencies
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.0.0"
ruff = "^0.1.0"
mypy = "^1.6.0"
pre-commit = "^3.5.0"

[tool.poetry.scripts]
price-alarm = "track:main"
setup-price-alarm = "setup:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"
